Make sure to only backup docker tor credentials if available

+++ Scripts/functions.sh
@@ -315,7 +315,7 @@ _is_active() {
 _tor_backup() {
     test -d "${tor_backup_dir}" || sudo mkdir -p "${tor_backup_dir}"
 
-    if [ -d "${dojo_path}" ]; then
+    if [ -d "${dojo_path}" ] && [ -d "${install_dir}/${tor_data_dir}"/_data/hsv3dojo ]; then
         sudo rsync -ac --delete-before --quiet "${install_dir}/${tor_data_dir}"/_data/ "${tor_backup_dir}"
         return 0
     fi
